requirements.txt
black
matplotlib
pillow
pre-commit
pytest
